+++ CHANGELOG.md
@@ -18,3 +18,12 @@
 ### 09 June, 2026
 
 - Added more sound effects.
+
+### 10 June, 2026
+
+- Added Guide screen for app usage tips.
+
+### 11 June, 2026
+
+- Fixed plan screen issues.
+- Converted sounds from wav to mp3.